sv: Change uiv_2buf to uivint_2buf
Printing out the full capability is unlikely to be desired. We also have
issues since TYPE_CHARS on a 128-bit data type gives an odd number of
chars due to rounding, unlike 32-bit and 64-bit ones, so if we want to
print out the full capability then we need to alter the start of the
function to do the last char on its own in that case before doing the
U16-at-a-time handling.