Use a string address rather than a string literal.
ENTER_with_name() stores a string literal address and LEAVE_with_name()
checks if the previously stored address is the same.

Explicitly use a string address to silence the warning:
warning: result of comparison against a string literal is unspecified
(use an explicit string comparison function instead)

Note that Perl assumes a compiler would maintain a pool of string
literals and use the same address for the same string literal in
calls to ENTER_with_name() and LEAVE_with_name(). This behaviour is
allowed but not required by the C language (footnote 86 in [0]).

Upstream [1] resolves this issue by ignoring -Wstring-compare for this
specific line.

[0] https://www.open-std.org/JTC1/SC22/WG14/www/docs/n1256.pdf
[1] https://github.com/Perl/perl5/commit/f7b332245abf1aa8ca9c82cce454353f7b820adb